* Do not try to check character case for sentences that have   been written fully in upper case. * Convert gc_sentence.token_count from int to size_t. * Improve error reporting for grammar check errors in voikkotest.
git-svn-id: https://voikko.svn.sourceforge.net/svnroot/voikko/trunk@2508 ee56993d-3414-0410-8ed4-8136e77a0dd8

+++ libvoikko/src/grammar/analysis.cpp
@@ -28,7 +28,7 @@ namespace libvoikko {
 /** Free the memory allocated for sentence analysis */
 void free_gc_sentence(gc_sentence * sentence) {
 	if (!sentence) return;
-	for (int i = 0; i < sentence->token_count; i++) {
+	for (size_t i = 0; i < sentence->token_count; i++) {
 		delete[] sentence->tokens[i].str;
 	}
 	delete sentence;
